cryptohack.org/01_General/01_Encoding/Bytes_and_Big_Integers.py
#!/usr/bin/env python3
import codecs
import Crypto.Util.number

the_int = 11515195063862318899931685488813747395775516287289682636499965282714637259206269

bytestr = Crypto.Util.number.long_to_bytes(the_int)
print(bytestr)
