cryptohack.org/requirements.txt
factordb-pycli
gmpy
pycryptodome
requests
pwntools
